Revert "Project restructure"
