+++ MatrixManager.cpp
