+++ MatrixManager.h
