+++ gui/ConfigureDialog.cpp
