+++ gui/ConfigureDialog.h
