+++ gui/LinearSystemDataEntry.cpp
