+++ gui/LinearSystemResultsDisplay.h
@@ -13,7 +13,7 @@
 #include <QTextEdit>
 
 #include "IntfResultsDisplay.h"
-#include "libmatrix/Matrix.h"
+#include "../Matrix.h"
 
 /**
  * @brief Layout for displaying a reduced linear system
