+++ gui/MainLayout.h
@@ -15,7 +15,7 @@
 #include "ConfigureDialog.h"
 #include "LinearSystemDataEntry.h"
 #include "LinearSystemResultsDisplay.h"
-#include "libmatrix/MatrixManager.h"
+#include "../MatrixManager.h"
 
 /**
  * @brief Layout instantiated in the main window
