+++ test/TestLinearSystem.h
@@ -7,8 +7,8 @@
 
 #include <random>
 #include <algorithm>
-#include "MatrixManager.h"
-#include "Matrix.h"
+#include "../MatrixManager.h"
+#include "../Matrix.h"
 
 #define LOG_FILE "./test_output.txt"
 class Random;
